+++ include/roq/api.h
@@ -17,12 +17,13 @@
 
 // enums
 
-#include "roq/cancel_policy.h"
 #include "roq/connection_status.h"
 #include "roq/error.h"
 #include "roq/execution_instruction.h"
 #include "roq/liquidity.h"
 #include "roq/option_type.h"
+#include "roq/order_cancel_policy.h"
+#include "roq/order_management.h"
 #include "roq/order_status.h"
 #include "roq/order_type.h"
 #include "roq/order_update_action.h"